+++ tests/conftest.py
@@ -176,3 +176,31 @@ def catalog(multi_backend_connection, config) -> AggregatorCollectionCatalog:
         backends=multi_backend_connection,
         config=config
     )
+
+
+JSON_CAPABILITIES_NO_ENDPOINTS = {
+    "api_version": "1.1.0",
+    "backend_version": "0.6.5a1.dev20221208+820",
+    "description": "openEO API for unit tests without secondary services",
+    "endpoints": [],
+    "id": "unittestopeneoapi-1.1.0",
+    "links": [],
+    "stac_version": "0.9.0",
+    "title": "Unit test API for OpenEO - no secondary services",
+    "version": "1.1.0",
+}
+
+
+JSON_CAPABILITIES_WITH_SERVICE_TYPES_SUPPORTED = {
+    "api_version": "1.1.0",
+    "backend_version": "0.6.5a1.dev20221208+820",
+    "description": "openEO API for unit tests with secondary services",
+    "endpoints": [
+        {"methods": ["GET"], "path": "/service_types"},
+    ],
+    "id": "unittestopeneoapi-1.1.0",
+    "links": [],
+    "stac_version": "0.9.0",
+    "title": "Unit test API for OpenEO - with secondary services",
+    "version": "1.1.0",
+}